fixed UI not updating, corrected Ganon's death event

+++ soh/include/z64save.h
@@ -223,7 +223,7 @@ typedef struct ShipQuestSaveContextData {
 } ShipQuestSaveContextData;
 
 typedef struct ShipAchievementSaveContextData {
-    u16 achievementFlags;
+    u32 achievementFlags;
 } ShipAchievementSaveContextData;
 
 typedef struct ShipSaveContextData {
